add create order integration

+++ src/main/java/com/cliente/rasmoo/plus/dtos/wsraspay/OrderDto.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class OrderDto {
     private String customerId;
-    private Long discount;
+    private BigDecimal discount;
     private String id;
     private String productAcronym;
 }
